requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
PyYAML==6.0.2
python-dateutil==2.9.0.post0
pydantic==2.9.2
imapclient==3.0.1
python-dotenv==1.0.1
